core/src/main/java/com/adobe/aem/guides/wknd/core/models/TestTwo.java
package com.adobe.aem.guides.wknd.core.models;

import org.apache.sling.api.resource.Resource;
import org.apache.sling.models.annotations.DefaultInjectionStrategy;
import org.apache.sling.models.annotations.Model;
import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;


@Model(adaptables = Resource.class, defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL)
public class TestTwo {
    @ValueMapValue
    private String name;

    @ValueMapValue
    private String location;

    public String getName() {
        return name;
    }

    public String getLocation() {
        return location;
    }
}